# specs/002-smart-note-management/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: MindNote API
  description: 智能笔记管理RESTful API，支持笔记CRUD、AI分析、搜索等功能
  version: 1.0.0
  contact:
    name: MindNote Team
    email: support@mindnote.app
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.mindnote.app/v1
    description: Production server
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          $ref: '#/components/schemas/ErrorResponseDetail'
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
        requestId:
          type: string
          example: 'req_123456789'
        version:
          type: string
          example: '1.0.0'
    ErrorResponseDetail:
      type: object
      properties:
        code:
          type: string
          example: 'VALIDATION_ERROR'
        message:
          type: string
          example: 'Request validation failed'
        details:
          type: object

    # 分页响应
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        total:
          type: integer
          example: 150
        totalPages:
          type: integer
          example: 8
        hasNext:
          type: boolean
          example: true
        hasPrev:
          type: boolean
          example: false

    # 核心实体
    Note:
      type: object
      properties:
        id:
          type: string
          example: 'note_123456789'
        title:
          type: string
          example: 'My Smart Note'
        content:
          type: string
          example: 'This is the content of my note...'
        contentHash:
          type: string
          example: 'abc123def456'
        categoryId:
          type: integer
          nullable: true
        category:
          $ref: '#/components/schemas/Category'
        tags:
          type: array
          items:
            $ref: '#/components/schemas/Tag'
        metadata:
          type: object
          example:
            source: 'web'
            priority: 'high'
        aiProcessed:
          type: boolean
          example: true
        aiSummary:
          type: string
          nullable: true
          example: 'A comprehensive note about productivity...'
        aiKeywords:
          type: array
          items:
            type: string
          example: ['productivity', 'tools', 'review']
        status:
          $ref: '#/components/schemas/NoteStatus'
        isPublic:
          type: boolean
          example: false
        viewCount:
          type: integer
          example: 15
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        aiProcessedAt:
          type: string
          format: date-time
          nullable: true
...
    NoteStatus:
      type: string
      enum:
        - DRAFT
        - PUBLISHED
        - ARCHIVED
    Category:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'Work'
        description:
          type: string
          nullable: true
        parentId:
          type: integer
          nullable: true
        color:
          type: string
          example: '#3B82F6'
        icon:
          type: string
          example: 'briefcase'
        noteCount:
          type: integer
          example: 25
        children:
          type: array
          items:
            $ref: '#/components/schemas/Category'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Tag:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: 'important'
        color:
          type: string
          example: '#EF4444'
        category:
          type: string
          example: 'priority'
        usageCount:
          type: integer
          example: 45
        createdAt:
          type: string
          format: date-time

    # AI相关
    AIAnalysisRequest:
      type: object
      required:
        - operations
      properties:
        operations:
          type: array
          items:
            type: string
            enum:
              - categorize
              - tag
              - summarize
          example: ['categorize', 'tag', 'summarize']
        provider:
          type: string
          example: 'auto'
        options:
          type: object
          properties:
            language:
              type: string
              example: 'zh-CN'
            summaryLength:
              type: integer
              example: 200
            maxTags:
              type: integer
              example: 5
            temperature:
              type: number
              example: 0.3
    AIAnalysisResponse:
      type: object
      properties:
        taskId:
          type: string
          example: 'ai_task_123456789'
        status:
          type: string
          enum:
            - QUEUED
            - PROCESSING
            - COMPLETED
            - FAILED
          example: 'COMPLETED'
        progress:
          type: integer
          example: 100
        result:
          type: object
          properties:
            category:
              type: string
              example: 'Work'
            tags:
              type: array
              items:
                type: string
              example: ['productivity', 'tools', 'review']
            summary:
              type: string
              example: 'A comprehensive review of productivity tools...'
            confidence:
              type: number
              example: 0.92
            provider:
              type: string
              example: 'zhipu'
            model:
              type: string
              example: 'glm-4'
            processingTime:
              type: integer
              example: 2450
            cost:
              type: number
              example: 0.025
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true

    # 搜索相关
    SearchRequest:
      type: object
      required:
        - query
      properties:
        query:
          type: string
          example: 'productivity tools'
        filters:
          type: object
          properties:
            categories:
              type: array
              items:
                type: integer
              example: [1, 2]
            tags:
              type: array
              items:
                type: string
              example: ['important']
            dateRange:
              type: object
              properties:
                from:
                  type: string
                  format: date
                  example: '2025-10-01'
                to:
                  type: string
                  format: date
                  example: '2025-10-31'
            status:
              type: array
              items:
                type: string
                enum:
                  - DRAFT
                  - PUBLISHED
                  - ARCHIVED
              example: ['PUBLISHED']
        sort:
          type: object
          properties:
            field:
              type: string
              enum:
                - relevance
                - createdAt
                - updatedAt
                - title
                - viewCount
              example: 'relevance'
            order:
              type: string
              enum:
                - asc
                - desc
              example: 'desc'
        pagination:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 20
    SearchResponse:
      type: object
      properties:
        items:
          type: array
          items:
            allOf:
              - $ref: '#/components/schemas/Note'
              - type: object
                properties:
                  relevanceScore:
                    type: number
                    example: 0.95
                  highlights:
                    type: array
                    items:
                      type: string
                    example: ['Latest <mark>productivity</mark> <mark>tools</mark>']
        pagination:
          $ref: '#/components/schemas/Pagination'
        searchMeta:
          type: object
          properties:
            queryTime:
              type: integer
              example: 45
            totalResults:
              type: integer
              example: 15
  parameters:
    NoteId:
      name: noteId
      in: path
      required: true
      schema:
        type: string
      description: Note ID
...
paths:
...
  /notes/{noteId}/ai-analyze:
    post:
      tags:
        - AI
      summary: Trigger AI analysis
      description: Start AI analysis for a note (categorization, tagging, summarization)
      parameters:
        - $ref: '#/components/parameters/NoteId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AIAnalysisRequest'
      responses:
        '200':
          description: AI analysis started
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/AIAnalysisResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '502':
          description: AI service unavailable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /search:
    post:
      tags:
        - Search
      summary: Full-text search
      description: Search notes with advanced filtering and ranking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /categories:
    get:
      tags:
        - Categories
      summary: Get categories
      description: Get user's categories with note counts and hierarchy
      parameters:
        - name: includeEmpty
          in: query
          schema:
            type: boolean
            default: false
          description: Include categories with no notes
      responses:
        '200':
          description: Categories retrieved
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Category'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Categories
      summary: Create category
      description: Create a new category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Work Projects'
                description:
                  type: string
                  nullable: true
                parentId:
                  type: integer
                  nullable: true
                color:
                  type: string
                  example: '#3B82F6'
                icon:
                  type: string
                  example: 'briefcase'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Category'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # 标签管理
...